Added adding client feature.
Added DialogForm, base class for future dialogs.
Added VDClientDetail, for adding or modifying client.
Modified money type float to decimal.
Updated foreign keys relation ships between product and
Sill/Time-interest.
Added IdCardNumber for Client class for make sure a Client is unique.

+++ OKInvestir/Model/Order.cs
@@ -17,7 +17,7 @@ namespace OKInvestir.Model
         [Required]
         public BoughtProduct BoughtProduct { get; set; }
         [Required]
-        public float FinalInterest { get; set; }
+        public decimal FinalInterest { get; set; }
         [Required]
         public DateTime Date { get; set; }
     }
